scraper/pyproject.toml
[project]
name = "chunimai-scraper"
version = "0.1.0"
description = "Daily play data scraper for Chunimai Dashboard"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "requests>=2.32",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
